requirements.txt
alembic==1.12.0
asyncpg==0.28.0
fastapi==0.103.2
psycopg2-binary==2.9.8
pydantic==2.4.2
pydantic-settings==2.0.3
uvicorn==0.23.2
SQLAlchemy==2.0.21
passlib==1.7.4
python-dotenv==1.0.1
python-jose==3.3.0
PyJWT==2.6.0